@@ -21,46 +21,47 @@
   "private": true,
   "packageManager": "pnpm@9.2.0",
   "dependencies": {
-    "@angular/animations": "^18.0.2",
-    "@angular/common": "^18.0.2",
-    "@angular/compiler": "^18.0.2",
-    "@angular/core": "^18.0.2",
-    "@angular/forms": "^18.0.2",
-    "@angular/platform-browser": "^18.0.2",
-    "@angular/platform-browser-dynamic": "^18.0.2",
-    "@angular/pwa": "^18.0.3",
-    "@angular/router": "^18.0.2",
-    "@angular/service-worker": "^18.0.2",
-    "@microsoft/signalr": "^8.0.0",
+    "@angular/animations": "^18.1.1",
+    "@angular/common": "^18.1.1",
+    "@angular/compiler": "^18.1.1",
+    "@angular/core": "^18.1.1",
+    "@angular/forms": "^18.1.1",
+    "@angular/platform-browser": "^18.1.1",
+    "@angular/platform-browser-dynamic": "^18.1.1",
+    "@angular/pwa": "^18.1.1",
+    "@angular/router": "^18.1.1",
+    "@angular/service-worker": "^18.1.1",
+    "@microsoft/signalr": "^8.0.7",
     "@ngneers/easy-ngrx-distinct-selector": "^0.1.1",
-    "@ngneers/signal-translate": "^0.1.1",
-    "@ngrx/effects": "18.0.0-rc.0",
-    "@ngrx/entity": "18.0.0-rc.0",
-    "@ngrx/operators": "18.0.0-rc.0",
-    "@ngrx/router-store": "18.0.0-rc.0",
-    "@ngrx/store": "18.0.0-rc.0",
-    "@ngrx/store-devtools": "18.0.0-rc.0",
+    "@ngneers/signal-translate": "^0.1.3",
+    "@ngrx/effects": "18.0.1",
+    "@ngrx/entity": "18.0.1",
+    "@ngrx/operators": "18.0.1",
+    "@ngrx/router-store": "18.0.1",
+    "@ngrx/store": "18.0.1",
+    "@ngrx/store-devtools": "18.0.1",
     "@types/eslint": "^8.56.10",
     "copy-to-clipboard": "^3.3.3",
-    "eruda": "^3.0.1",
+    "eruda": "^3.2.1",
     "immer": "^10.1.1",
-    "primeng": "17.18.0",
+    "primeng": "17.18.5",
     "rxjs": "~7.8.1",
     "tslib": "^2.6.3",
-    "zone.js": "~0.14.7"
+    "tsx": "^4.16.2",
+    "zone.js": "~0.14.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.0.3",
-    "@angular-eslint/builder": "^18.0.1",
-    "@angular-eslint/eslint-plugin": "^18.0.1",
-    "@angular-eslint/eslint-plugin-template": "^18.0.1",
-    "@angular-eslint/template-parser": "^18.0.1",
-    "@angular/cli": "^18.0.3",
-    "@angular/compiler-cli": "^18.0.2",
-    "@goast/core": "^0.2.1",
-    "@goast/typescript": "^0.2.15",
+    "@angular-devkit/build-angular": "^18.1.1",
+    "@angular-eslint/builder": "^18.1.0",
+    "@angular-eslint/eslint-plugin": "^18.1.0",
+    "@angular-eslint/eslint-plugin-template": "^18.1.0",
+    "@angular-eslint/template-parser": "^18.1.0",
+    "@angular/cli": "^18.1.1",
+    "@angular/compiler-cli": "^18.1.1",
+    "@goast/core": "^0.3.1",
+    "@goast/typescript": "^0.3.1",
     "@iconify-json/flag": "^1.1.53",
-    "@iconify-json/mdi": "^1.1.66",
+    "@iconify-json/mdi": "^1.1.67",
     "@iconify/tailwind": "^1.1.1",
     "@ngneers/eslint-config": "^1.0.0",
     "@ngneers/eslint-config-angular": "^1.0.1",
@@ -80,13 +81,13 @@
     "http-server": "^14.1.1",
     "jest": "^29.7.0",
     "jest-junit": "^16.0.0",
-    "jest-preset-angular": "^14.1.0",
-    "postcss": "^8.4.38",
-    "prettier": "^3.3.1",
-    "prettier-plugin-tailwindcss": "^0.6.2",
+    "jest-preset-angular": "^14.1.1",
+    "postcss": "^8.4.39",
+    "prettier": "^3.3.3",
+    "prettier-plugin-tailwindcss": "^0.6.5",
     "run-script-os": "^1.1.6",
-    "tailwindcss": "^3.4.4",
-    "ts-jest": "^29.1.4",
-    "typescript": "~5.4.5"
+    "tailwindcss": "^3.4.6",
+    "ts-jest": "^29.2.3",
+    "typescript": "~5.5.3"
   }
 }
